requirements.txt
# 多Agent协同运营自动化系统 - Python依赖
# 核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 异步消息队列
redis==5.0.1
aioredis==2.0.1

# 数据库
aiosqlite==0.19.0
sqlalchemy==2.0.25

# 任务调度
apscheduler==3.10.4

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# LLM支持
openai==1.10.0
tiktoken==0.5.2

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 工具
python-multipart==0.0.6
email-validator==2.1.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# 开发
black==24.1.1
ruff==0.1.14
